chore: resolve the Rider inspection list where the compiler agrees
199 of 214 listed sites: dead parameters/locals/usings, doc crefs to runtime types the generator cannot
reference, redundant casts and suppressions, static local functions, multiple enumeration in the graph
oracle, a bit-exact float comparison in the AOT bench, and a stated-reason suppression for every deliberate
name, namespace and public-surface null guard. Left alone on purpose: Rider's nullability guesses in the
netstandard2.0 projects, three analyzer/test-contract conflicts, hook-contract parameters (ledger P11).

+++ tests/DwarfMapper.IntegrationTests/InstructionSetMatrixTests.cs
@@ -129,7 +129,7 @@ namespace DwarfMapper.IntegrationTests
             for (var i = 0; i < n; i++)
             {
                 // The scalar oracle, computed here rather than trusted from the mapper.
-                Assert.Equal((long)src[i], dst.V[i]);
+                Assert.Equal(src[i], dst.V[i]);
             }
         }
 
@@ -140,7 +140,7 @@ namespace DwarfMapper.IntegrationTests
 
             var dst = new IsaMapper().MapWiden(new IsaWidenSrc { V = src });
 
-            Assert.Equal([(long)int.MinValue, -1L, 0L, 1L, int.MaxValue], dst.V);
+            Assert.Equal([int.MinValue, -1L, 0L, 1L, int.MaxValue], dst.V);
         }
 
         [Theory]
